Apply multiple reductions in tests if appropriate
Replace `(label_of (a ->)) ->` with `(label_of ((a ->)->)) ->`
or `(label_of (((a ->)->)->) ->` to restore coverage of appropriate
tests.

Most of the reduction tests were written when `a ->` used to mean
"print the normal form of `a`". When the language design was changed
so that `a ->` meant "the concept that `a` reduces to", the tests that
tried to print the normal form of a concept instead tested printing
the reduction of a concept. They should test printing of normal form.