Reveal problem in calling concept_from_ast method
The concept_from_ast method in the ConceptMaker trait assumes if
a `SyntaxTree` has its `concept` field equal to None then the context
has no concept that is labelled with its syntax. The problem is that
concept_from_ast is called sequentially in `ExecuteReduction` trait's
method. This causes the arguments to concept_from_ast to be out of sync
with the context.