symbols of form _(.(?!\s))*_ noted as variables
The strings are not saved to Context after execution.
However the concept of the label still persists. Need to find a
clean way to temporarily store variable symbols so that the same
variable can be referred to in the same expression but that doesn't
interfer with other functionality.

The test in the examples that is failing can pass if the reduce method
is modified to search for generalisations of the syntax tree if a
reduction cannot be found for the particular syntax tree