:recycle: Restructure Concept preventing need to iterate
Instead of storing a HashSet of concept IDs that the Concept is the lefthand or righthand of, the same concept IDs can be stored as values of a HashMap with the key being the other part of the composition. This prevents the need to iterate over a hashset in multiple methods which would cause performance issues for concepts that are used in many compositions. The use of Option::and_then was also replaced with the try operator, '?', where applicable to reduce indentation