:recycle: Use slotmap instead of vector of options
Ended up defining and implementing ConceptTrait for Concept<Id> and a new enum, Mixed<'a> that deals with the case of a previously committed concept with some uncommitted changes and a concept which is itself uncommitted. These abstractions prevent having to clone whole concepts just to read a property of a concept. It also paves the way for a efficient way to load concepts from a database.